requirements.txt
bentoml
scikit-learn
pandas
numpy
pydantic
starlette
pyjwt
requests

